+++ modules/parameter/controllers/UserController.php
@@ -37,7 +37,10 @@ class UserController extends Controller
      * @return mixed
      */
     public function actionIndex()
-    {    
+    {
+        if(!$this->cekAkses()){          
+            throw new NotFoundHttpException('You don\'t have access.');
+        }
         $searchModel = new UserSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -54,7 +57,10 @@ class UserController extends Controller
      * @return mixed
      */
     public function actionView($id)
-    {   
+    {
+        if(!$this->cekAkses()){          
+            throw new NotFoundHttpException('You don\'t have access.');
+        }
         $request = Yii::$app->request;
         if($request->isAjax){
             Yii::$app->response->format = Response::FORMAT_JSON;
@@ -81,6 +87,9 @@ class UserController extends Controller
      */
     public function actionCreate()
     {
+        if(!$this->cekAkses()){          
+            throw new NotFoundHttpException('You don\'t have access.');
+        }
         $request = Yii::$app->request;
         $model = new User();  
 
@@ -151,6 +160,9 @@ class UserController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(!$this->cekAkses()){          
+            throw new NotFoundHttpException('You don\'t have access.');
+        }
         $request = Yii::$app->request;
         $model = $this->findModel($id);       
 
@@ -219,6 +231,9 @@ class UserController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!$this->cekAkses()){          
+            throw new NotFoundHttpException('You don\'t have access.');
+        }
         $request = Yii::$app->request;
         $this->findModel($id)->delete();
 
@@ -246,7 +261,10 @@ class UserController extends Controller
      * @return mixed
      */
     public function actionBulkDelete()
-    {        
+    {    
+        if(!$this->cekAkses()){          
+            throw new NotFoundHttpException('You don\'t have access.');
+        }    
         $request = Yii::$app->request;
         $pks = explode(',', $request->post( 'pks' )); // Array or selected records primary keys
         foreach ( $pks as $pk ) {
@@ -282,6 +300,15 @@ class UserController extends Controller
             return $model;
         } else {
             throw new NotFoundHttpException('The requested page does not exist.');
+        }
+    }
+
+    // cek akses user
+    protected function cekAkses(){
+        if(!Yii::$app->user->isGuest && Yii::$app->user->identity->group_id <= 1){
+            return true;
+        }ELSE{
+            return false;
         }
     }
 }
